requirements.txt
google-genai==1.60.0
markdown==3.10.1
python-dotenv==1.2.1